+++ internal/component/common/kubernetes/mimir_rule_group_diff_test.go
@@ -4,30 +4,31 @@ import (
 	"fmt"
 	"testing"
 
-	"github.com/prometheus/prometheus/model/rulefmt"
+	"github.com/grafana/alloy/internal/mimir/client"
 	"github.com/stretchr/testify/require"
 )
 
-func parseRuleGroups(t *testing.T, buf []byte) []rulefmt.RuleGroup {
+func parseMimirRuleGroups(t *testing.T, buf []byte) []client.MimirRuleGroup {
 	t.Helper()
 
-	groups, errs := rulefmt.Parse(buf)
+	groups, errs := client.Parse(buf)
 	require.Empty(t, errs)
 
 	return groups.Groups
 }
 
-func TestDiffRuleState(t *testing.T) {
-	ruleGroupsA := parseRuleGroups(t, []byte(`
+func TestDiffMimirRuleGroupState(t *testing.T) {
+	ruleGroupsA := parseMimirRuleGroups(t, []byte(`
 groups:
 - name: rule-group-a
   interval: 1m
+  source_tenants: ["tenant1","tenant2"]
   rules:
   - record: rule_a
     expr: 1
 `))
 
-	ruleGroupsAModified := parseRuleGroups(t, []byte(`
+	ruleGroupsAModified := parseMimirRuleGroups(t, []byte(`
 groups:
 - name: rule-group-a
   interval: 1m
@@ -40,25 +41,25 @@ groups:
 
 	type testCase struct {
 		name     string
-		desired  map[string][]rulefmt.RuleGroup
-		actual   map[string][]rulefmt.RuleGroup
-		expected map[string][]RuleGroupDiff
+		desired  map[string][]client.MimirRuleGroup
+		actual   map[string][]client.MimirRuleGroup
+		expected map[string][]MimirRuleGroupDiff
 	}
 
 	testCases := []testCase{
 		{
 			name:     "empty sets",
-			desired:  map[string][]rulefmt.RuleGroup{},
-			actual:   map[string][]rulefmt.RuleGroup{},
-			expected: map[string][]RuleGroupDiff{},
+			desired:  map[string][]client.MimirRuleGroup{},
+			actual:   map[string][]client.MimirRuleGroup{},
+			expected: map[string][]MimirRuleGroupDiff{},
 		},
 		{
 			name: "add rule group",
-			desired: map[string][]rulefmt.RuleGroup{
+			desired: map[string][]client.MimirRuleGroup{
 				managedNamespace: ruleGroupsA,
 			},
-			actual: map[string][]rulefmt.RuleGroup{},
-			expected: map[string][]RuleGroupDiff{
+			actual: map[string][]client.MimirRuleGroup{},
+			expected: map[string][]MimirRuleGroupDiff{
 				managedNamespace: {
 					{
 						Kind:    RuleGroupDiffKindAdd,
@@ -69,11 +70,11 @@ groups:
 		},
 		{
 			name:    "remove rule group",
-			desired: map[string][]rulefmt.RuleGroup{},
-			actual: map[string][]rulefmt.RuleGroup{
+			desired: map[string][]client.MimirRuleGroup{},
+			actual: map[string][]client.MimirRuleGroup{
 				managedNamespace: ruleGroupsA,
 			},
-			expected: map[string][]RuleGroupDiff{
+			expected: map[string][]MimirRuleGroupDiff{
 				managedNamespace: {
 					{
 						Kind:   RuleGroupDiffKindRemove,
@@ -84,13 +85,13 @@ groups:
 		},
 		{
 			name: "update rule group",
-			desired: map[string][]rulefmt.RuleGroup{
+			desired: map[string][]client.MimirRuleGroup{
 				managedNamespace: ruleGroupsA,
 			},
-			actual: map[string][]rulefmt.RuleGroup{
+			actual: map[string][]client.MimirRuleGroup{
 				managedNamespace: ruleGroupsAModified,
 			},
-			expected: map[string][]RuleGroupDiff{
+			expected: map[string][]MimirRuleGroupDiff{
 				managedNamespace: {
 					{
 						Kind:    RuleGroupDiffKindUpdate,
@@ -102,28 +103,28 @@ groups:
 		},
 		{
 			name: "unchanged rule groups",
-			desired: map[string][]rulefmt.RuleGroup{
+			desired: map[string][]client.MimirRuleGroup{
 				managedNamespace: ruleGroupsA,
 			},
-			actual: map[string][]rulefmt.RuleGroup{
+			actual: map[string][]client.MimirRuleGroup{
 				managedNamespace: ruleGroupsA,
 			},
-			expected: map[string][]RuleGroupDiff{},
+			expected: map[string][]MimirRuleGroupDiff{},
 		},
 	}
 
 	for _, tc := range testCases {
 		t.Run(tc.name, func(t *testing.T) {
-			actual := DiffRuleState(tc.desired, tc.actual)
-			requireEqualRuleDiffs(t, tc.expected, actual)
+			actual := DiffMimirRuleGroupState(tc.desired, tc.actual)
+			requireEqualMimirRuleGroupDiffs(t, tc.expected, actual)
 		})
 	}
 }
 
-func requireEqualRuleDiffs(t *testing.T, expected, actual map[string][]RuleGroupDiff) {
+func requireEqualMimirRuleGroupDiffs(t *testing.T, expected, actual map[string][]MimirRuleGroupDiff) {
 	require.Equal(t, len(expected), len(actual))
 
-	var summarizeDiff = func(diff RuleGroupDiff) string {
+	var summarizeDiff = func(diff MimirRuleGroupDiff) string {
 		switch diff.Kind {
 		case RuleGroupDiffKindAdd:
 			return fmt.Sprintf("add: %s", diff.Desired.Name)
@@ -145,8 +146,8 @@ func requireEqualRuleDiffs(t *testing.T, expected, actual map[string][]RuleGroup
 			actualDiff := actualDiffs[i]
 
 			if expectedDiff.Kind != actualDiff.Kind ||
-				!equalRuleGroups(expectedDiff.Desired, actualDiff.Desired) ||
-				!equalRuleGroups(expectedDiff.Actual, actualDiff.Actual) {
+				!equalMimirRuleGroups(expectedDiff.Desired, actualDiff.Desired) ||
+				!equalMimirRuleGroups(expectedDiff.Actual, actualDiff.Actual) {
 
 				t.Logf("expected diff: %s", summarizeDiff(expectedDiff))
 				t.Logf("actual diff: %s", summarizeDiff(actualDiff))